Support decimal prices (cents) end-to-end (#21) (#22)
Migrate price columns from BIGINT to NUMERIC(20,2), update Go models
from int64 to float64, and make the frontend price input preserve
decimals so pasting "19090.00" from Janice works correctly.

+++ internal/controllers/buyOrders.go
@@ -72,7 +72,7 @@ func (c *BuyOrdersController) CreateOrder(args *web.HandlerArgs) (any, *web.Http
 	var req struct {
 		TypeID          int64   `json:"typeId"`
 		QuantityDesired int64   `json:"quantityDesired"`
-		MaxPricePerUnit int64   `json:"maxPricePerUnit"`
+		MaxPricePerUnit float64 `json:"maxPricePerUnit"`
 		Notes           *string `json:"notes"`
 	}
 
@@ -161,7 +161,7 @@ func (c *BuyOrdersController) UpdateOrder(args *web.HandlerArgs) (any, *web.Http
 
 	var req struct {
 		QuantityDesired int64   `json:"quantityDesired"`
-		MaxPricePerUnit int64   `json:"maxPricePerUnit"`
+		MaxPricePerUnit float64 `json:"maxPricePerUnit"`
 		Notes           *string `json:"notes"`
 		IsActive        *bool   `json:"isActive"`
 	}
